MacOS: Fix invisible PopupMenu in Logic
We avoid modifying window controls in modal situations if possible.
Doing as much as [window setStyleMask: [window styleMask]]; for a
PopupMenu window will make it invisible on MacOS when the plugin is
hosted in an external subprocess.

This fixes a regression caused by
fb14118771eeb34f51146e23def0a1f94a49fbad.